% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rf_eq_estimate}
\alias{rf_eq_estimate}
\title{Distributionally Adaptive Regression Estimator via Random Forest}
\usage{
rf_eq_estimate(Y, X, x0, nsize = 20, c_band = 1, K_grid, L = 5, ntr = 1000L)
}
\arguments{
\item{Y}{Numeric vector length n.}

\item{X}{Numeric matrix/data.frame n x p.}

\item{x0}{Numeric length-p vector or 1-row matrix for evaluation.}

\item{nsize}{Integer, nodesize for quantregForest. Default 20.}

\item{c_band}{Numeric bandwidth multiplier. Default 1.}

\item{K_grid}{Integer vector of candidate K values.}

\item{L}{Integer, number of CV folds. Default 5.}

\item{ntr}{Integer, number of trees. Default 1000.}
}
\value{
Numeric scalar estimate m(x0).
}
\description{
Distributionally Adaptive Regression Estimator via Random Forest
}
